refactor: Update import paths and references from utilization-metering-consumer to event-router
Update all Go import paths, Kafka client ID, consumer group ID default,
service name constants, Dockerfile binary name and build path, and
package-level comments to reflect the new event-router service name.
